[SUMMARY] Implemented the hybridizationPercentBound function. Awaiting results from team members to see where improvements could be had.
Hoping for Glorious Amazing God Tier results. Likely awaiting Shitty Terrible News Your Cat Is A Frog.

Five more pages to go in document, but likely notes and explanations that are handy but not useful for the programmers.

Will attempt to bug Wilson with use cases, and likely attempt to use the IDT calculator to test things out.